requirements.txt
streamlit==1.31.0
pandas==2.2.0
openai==1.12.0
python-dotenv==1.0.1
selenium==4.16.0
webdriver-manager==4.0.1
scikit-learn==1.3.2
numpy==1.26.3
supabase==2.3.4
beautifulsoup4==4.12.3
requests==2.31.0 